backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
firebase-admin>=6.4,<7.0
PyJWT>=2.8,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
